Applr: Tools for Visualizing and Explaining Linear Models

Applr is a small R package that makes it easy to visualize slices of higher-dimensional linear models and to present model equations clearly. It includes base R and ggplot2 helpers, plus an interactive 3D view for models with two predictors.

Details

Core capabilities:

2D slices of multi-variable models in base graphics

ggplot2 layer for plotting model slices across facets

Interactive 3D scatter with fitted surface (two predictors)

Quick text/LaTeX equations from lm() objects

Main functions

autoplot: autoplot.lm

Base graphics: slice_2d, add_slice_2d

ggplot2 layer: geom_slice (built on StatSlice and GeomSlice)

ggplot2 labeling: geom_slice_text, geom_slice_subtitle, geom_slice_caption

3D visualization: scatter_3d

Equations: lm_equation, lm_latex

Theme: theme_lc

Quick start

Fit a model and plot a slice:

library(Applr)
model <- lm(mpg ~ disp + hp, data = mtcars)
slice_2d(model, x_axis = "disp")

ggplot2 layer with facets:

library(ggplot2)
model <- lm(mpg ~ disp + hp + cyl, data = mtcars)
ggplot(mtcars, aes(disp, mpg)) +
  geom_point() +
  facet_wrap(~cyl) +
  geom_slice(model = model, predict_vars = list(hp = 110))

See also

Author

Maintainer: James Beeson jamesbeeson01@gmail.com

Authors:

Cameron McClellan cameron.m.mcclellan@gmail.com